refactor normalization core records
